Bump datatype of item flags from u16 to u32  (#3792)

+++ Database/itemDB.lua
@@ -89,7 +89,7 @@ end
 QuestieDB.itemCompilerTypes = {
     ["foodType"] = "u8",
     ["itemLevel"] = "u8",
-    ["flags"] = "u16",
+    ["flags"] = "u32",
     ["startQuest"] = "u24",
     ["requiredLevel"] = "u8",
     ["ammoType"] = "u8",
